requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.0.1
pytesseract==0.3.10
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-WTF==1.2.2